Refactoring - break out to_s-type methods into presenters
