--- app/presenters/time_data/base_presenter.rb
@@ -1,7 +1,0 @@
-module TimeData
-
-  class BasePresenter < SimpleDelegator
-
-  end
-
-end
